+++ Makefile
@@ -1,7 +1,7 @@
 # $PostgreSQL: pgsql/contrib/plpgsql_check/Makefile
 
 MODULE_big = plpgsql_check
-OBJS = plpgsql_check.o
+OBJS = format.o catalog.o tablefunc.o plpgsql_setup.o expr_walk.o assign.o check_expr.o report.o stmtwalk.o
 DATA =  plpgsql_check--1.0--1.1.sql plpgsql_check--1.4.sql plpgsql_check--1.1--1.2.sql plpgsql_check--1.2--1.3.sql plpgsql_check--1.3--1.4.sql
 EXTENSION = plpgsql_check
